User Profile Page - major update
Core design for the user profile show view updated. Ended up sticking
to a two-column layout with photo and user info to the right of a main
content div that includes the user's bio. Thinking of a way to spice up
the main header, or perhaps reduce it's height…will revisit.