Directory Page - Bux Fix, Images Reformatted
Had an issue on the directory page that was breaking the layout and
design, had to do with pulling images from LinkedIn that were not
formatted as .jpg and were being dynamically changed as car as
dimension. PS couldn't read also - this only happens for images that
don't expand. Resaved, problem fixed.